requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dateutil==2.9.0.post0
pytest==8.3.3
pytest-asyncio==0.24.0
anyio==4.4.0
